Add function to play a round of rps
If the player choice is the same as computer choice, it's a tie
and we are done. Otherwise, we enter the switch statement to check
for each case.

Within the switch statement, I did not explicitly write conditionals
for the "lose" outcome because if the player does not win against
the computer, the only other outcome is "lose"; if it were a tie, we
would have never even entered the switch statement.