Add counters to keep track of score and declare winner
I am manually changing the value of the iteration variable i
within the loop. If it's a tie, the loop decrements and the round
"does not count". If someone achieves the winning score, which in
this case is a score of 3 in a best of 5, setting i to the end of
the loop exits the loop and ends the game.