Refactor playRound function
Refactor playRound function to only return results of each round.
Since we got rid of the console.logs, we update the output with the
results of each round with the new updatesResultsDisplay function,
which also increments and updates the score variables.

The updateRoundDisplay clears the display in between rounds, but
also checks whether a winner was decided from the previous round
based on the scores, to decide if the scores should be reset to
default for an entire new match. The updateWinnerDisplay is just
something extra to add "special text" for when a winner of the
match is decided.

innerText property is used over textContent in some cases to create
line breaks.